% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retireving_functions.R
\name{geony_address}
\alias{geony_address}
\title{Extract longlat for a given address from geoclient v1 with a single string address}
\usage{
geony_address(address, app_id, key, simplify = T)
}
\arguments{
\item{address}{a single string address input, spaces are fine.}

\item{app_id}{app id.}

\item{key}{key.}
}
\description{
Extract longlat for a given address from geoclient v1 with a single string address
}
\examples{
geony_address("314west100stmanhattan", "f2e57cb4", "68779dfdbac8b55ecbe179e6d64dd734")
}
\keyword{geony}

