% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retireving_functions.R
\name{geony_all}
\alias{geony_all}
\title{Extracting all Data from geoclient V1 with a full partitioned address}
\usage{
geony_all(housenum, street, borough, app_id, key)
}
\arguments{
\item{housenum}{the house number.}

\item{street}{the street number.}

\item{borough}{borough.}

\item{app_id}{app id.}

\item{key}{key.}
}
\description{
This function returns all raw data from geo client for a given structured address
}
\examples{
geony_all("5123", "108st", "queens", "f2e57cb4", "68779dfdbac8b55ecbe179e6d64dd734")
}
\keyword{geony}

