% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retireving_functions.R
\name{geony_fips}
\alias{geony_fips}
\title{This function returns full fips code for a given long lat from the FCC API}
\usage{
geony_fips(latitude, longitude)
}
\arguments{
\item{housenum}{longitude.}

\item{street}{latitude.}
}
\description{
This function returns full fips code for a given long lat from the FCC API
}
\examples{
geony_fips(40.75, -73.9)
}
\keyword{geony}

