% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retireving_functions.R
\name{geony_intersect}
\alias{geony_intersect}
\title{Extract longlat for a given cross street}
\usage{
geony_intersect(cross_street_one, cross_street_two, borough, app_id, key,
  simplify = T)
}
\arguments{
\item{cross_street_one}{a single string cross street input, spaces are fine.}

\item{cross_street_two}{a single string cross street input, spaces are fine.}

\item{borough}{a borough input.}

\item{app_id}{app id.}

\item{key}{key.}
}
\description{
Extract longlat for a given cross street
}
\examples{
geony_intersect("broad st","beaver st","manhattan" ,"f2e57cb4", "68779dfdbac8b55ecbe179e6d64dd734")
}
\keyword{geony}

