% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retireving_functions.R
\name{geony_latlong}
\alias{geony_latlong}
\title{Extracting longlat for a given address from geoclient V1 with a full partitioned address}
\usage{
geony_latlong(housenum, street, borough, app_id, key, simplify = F)
}
\arguments{
\item{housenum}{the house number.}

\item{street}{the street number.}

\item{borough}{borough.}

\item{app_id}{app id.}

\item{key}{key.}
}
\description{
This function returns longlat from geo client for a given address structured
}
\examples{
geony_latlong("5123", "108st", "queens", "f2e57cb4", "68779dfdbac8b55ecbe179e6d64dd734")
}
\keyword{geony}

