% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retireving_functions.R
\name{geony_raw}
\alias{geony_raw}
\title{Extract longlat for a given address from geoclient v1 with a single string address}
\usage{
geony_raw(address, app_id, key, simplify = F)
}
\arguments{
\item{address}{a single string address input, spaces are fine.}

\item{app_id}{app id.}

\item{key}{key.}
}
\description{
Extract longlat for a given address from geoclient v1 with a single string address
}
\examples{
geony_raw("314west100stmanhattan", "f2e57cb4", "68779dfdbac8b55ecbe179e6d64dd734")
Unstructured query takes a multitude of parameters e.g (addresses, bbls, intersections) see https://api.cityofnewyork.us/geoclient/v1/doc#section-1.3 for for information.
}
\keyword{geony}

